+++ src/js/main/appReady.js
@@ -216,7 +216,7 @@ function appReady(args) {
     const options = {
         backgroundColor: args.transparent ? '#0fff' : '#fff',
         show: false,
-        frame: !(args.transparent || args['cover-displays']),
+        frame: args.frame && !args['cover-displays'],
         titleBarStyle: 'hidden',
         fullscreenWindowTitle: true,
         fullscreenable: true,