+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- add support for relative paths to `--serve`
+
 ## [0.14.1] 2020-12-02
 
 ### Fixed