+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - add support for relative paths to `--serve`
 - add `--resize`/`--no-resize` cli option to allow/prevent browser window resizing
+- add `--frame`/`--no-frame` cli option to show/hide browser window frame
 
 ## [0.14.1] 2020-12-02